requirements.txt
streamlit>=1.30.0
pandas>=1.5.3
numpy>=1.23.0
matplotlib>=3.7.0
plotly>=5.16.0
Pillow>=9.5.0
streamlit-drawable-canvas>=0.8.0
streamlit-image-coordinates>=0.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
